Feature/general optimizations

+++ world-composer/src/data/engine/EngineUtils.ts
@@ -55,7 +55,7 @@ export function calculateFieldPressure(
     );
 
     let iFluidCellInd = 0;
-    const globalPressureVector = new Float64Array(grid.size ** 2);
+    const globalPressureVector = new Float32Array(grid.size ** 2);
     globalPressureVector.forEach((_, ind, pressure) => {
         pressure[ind] =
             grid.solids[ind] === 1
